pc/requirements.txt
pystray>=0.19.0
Pillow>=10.0.0
websockets>=12.0
pyautogui>=0.9.54
pyperclip>=1.8.0
pyinstaller>=6.0.0
